[GGS Branch] Memory requirements estimation now considers render scale.

+++ branches/ggs/src/GeoGen/Configuration.hpp
@@ -12,7 +12,7 @@ namespace geogen
 		bool MainMapIsMandatory = true;
 
 		/// Maximum sum of memory footprints of all the maps allocated simultaneously by the Renderer, in bytes. Default: 100 MiB.
-		unsigned long long RendererMemoryLimit = 100 * 1024 * 1024;
+		unsigned long RendererMemoryLimit = 100 * 1024 * 1024;
 
 		virtual void Serialize(IOStream& stream) const;
 	};
